Refactor transformers
