+++ .gitignore
@@ -10,3 +10,4 @@ spark-warehouse/*
 derby.log
 tests/__pycache__/*
 schema/__pycache__/*
+transform/__pycache__/*
