+++ transform/PartiesTransformer.py
@@ -1,10 +1,10 @@
 from pyspark.sql.dataframe import DataFrame
 from pyspark.sql.functions import col
 
-from lib.Transformer import Transformer
+from transform.Transformer import Transformer
 
 
-class PartiesToRelationsTransformer(Transformer):
+class PartiesTransformer(Transformer):
 
     def transform(self, df: DataFrame) -> DataFrame:
         return df.select(
